added QR Popup Modal to Details Page using Global State Actions
